CNDB-12503 move tests into existing create index test file
